requirements.txt
Pillow
tqdm
numpy
torch
transformers
open_clip_torch
requests